// screen/src/types/enums.ts
export enum EventStatus {
  STANDBY = 'STATUS_STANDBY',
  CHECKIN = 'STATUS_CHECKIN',
  LOTTERY_READY = 'STATUS_LOTTERY_READY',
  LOTTERY_RUNNING = 'STATUS_LOTTERY_RUNNING',
  GAME_SHAKE = 'STATUS_GAME_SHAKE',
  ENDED = 'STATUS_ENDED',
}

export enum WsEvent {
  HOST_CHANGE_SCENE = 'HOST_CHANGE_SCENE',
  USER_SHAKE_ACTION = 'USER_SHAKE_ACTION',
  JOIN_ROOM = 'JOIN_ROOM',
  SCENE_UPDATED = 'SCENE_UPDATED',
  LOTTERY_WINNER_ANNOUNCE = 'LOTTERY_WINNER_ANNOUNCE',
  USER_CHECKED_IN = 'USER_CHECKED_IN',
  SHAKE_LEADERBOARD_TICK = 'SHAKE_LEADERBOARD_TICK',
}